+++ app/Http/Controllers/Submission/MomTaskUpdateController.php
@@ -4,22 +4,28 @@ namespace App\Http\Controllers\Submission;
 
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Mail;
+use App\Mail\SubmissionMail;
 
 use App\Models\Module;
 use App\Models\Submission\MomTaskUpdate;
+use App\Models\Submission\MomTaskBound;
+use App\Models\User;
+use DB;
 
 class MomTaskUpdateController extends Controller
 {
 
     private $model;
     public $modulename;
-    public $module;
+    public $taskbound;
 
     public function __construct()
     {
         $this->model = new MomTaskUpdate();
         $this->modulename = 'Mom';
         $this->module = new Module();
+        $this->taskbound = new MomTaskBound();
     }
 
     public function index()
@@ -43,8 +49,19 @@ class MomTaskUpdateController extends Controller
             $requestData = $request->all();
             $requestData['module_id'] = $this->getModuleId($request->modulename);
             $requestData['updated_by'] = $this->getEmployeeID()->id;
+            $requestData['date'] = date('Y-m-d');
 
-            $this->addOneDayToDate($requestData);
+            $getTaskBound = $this->taskbound->where('task_id',$request->task_id)
+            ->where('employee_id',$this->getEmployeeID()->id)
+            ->get();
+
+            if(count($getTaskBound) < 1) {
+                return response()->json(["status" => "error", "message" => $this->getMessage()['nothaveaccess']]);
+            }
+
+            // START NORIFICATION
+            $this->generateNotificationMessage($request->task_id, $mode = 'Add');
+            // END NORIFICATION
 
             $this->model->create($requestData);
 
@@ -82,13 +99,24 @@ class MomTaskUpdateController extends Controller
     public function update(Request $request, $id)
     {
         try {
+            $data = $this->model->findOrFail($id);
             
             $requestData = $request->all();
             $requestData['updated_by'] = $this->getEmployeeID()->id;
+            $requestData['date'] = date('Y-m-d');
 
-            $this->addOneDayToDate($requestData);
+            $getTaskBound = $this->taskbound->where('task_id',$data->task_id)
+            ->where('employee_id',$this->getEmployeeID()->id)
+            ->get();
 
-            $data = $this->model->findOrFail($id);
+            if(count($getTaskBound) < 1) {
+                return response()->json(["status" => "error", "message" => $this->getMessage()['nothaveaccess']]);
+            }
+            
+            // START NORIFICATION
+                $this->generateNotificationMessage($data->task_id, $mode = 'Update');
+            // END NORIFICATION
+
             $data->update($requestData);
 
             return response()->json(["status" => "success", "message" => $this->getMessage()['update']]);
@@ -99,11 +127,81 @@ class MomTaskUpdateController extends Controller
         }
     }
 
+    public function generateNotificationMessage($id, $mode) {
+        $getMomID = DB::table('request_momTask')->select('tbl_category.req_id')
+                        ->leftJoin('tbl_category','request_momTask.category_id','tbl_category.id')
+                        ->where('request_momTask.id',$id)
+                        ->first();
+        $getSubmissionData = DB::table('request_mom')->where('id', $getMomID->req_id)->first();
+        $getCreator = User::findOrFail($getSubmissionData->user_id); //  get creator
+
+        $mailData = [
+            "all" => 1,
+            "action_id" => 0,
+            "submission" => $getSubmissionData,
+            "email" => $getCreator->email, // kirim kepada creator
+            "fullname" => $getCreator->fullname,
+            "message" => $this->mailMessage()['newActivity'],
+            "remarks" => $mode,
+        ];
+        if($getSubmissionData->requestStatus == 3) {
+            Mail::to($mailData['email'])->send(new SubmissionMail($mailData,$this->modulename,1));
+        }
+    }
+
+    public function reminderNotificationMessage($mode) {
+        if($mode == 'reminder') {
+            
+            $reminders = [];
+            $getReminderMom = DB::table('reminderMomDeadline')->get();
+            foreach($getReminderMom as $r) {
+                $reminders[] = $r->mom_id;
+            }
+            $uniqueReminders = array_unique($reminders);
+
+            // return $uniqueReminders;
+
+            foreach($uniqueReminders as $g) {
+
+                // $getMomID = DB::table('request_momTask')->select('tbl_category.req_id')
+                //                 ->leftJoin('tbl_category','request_momTask.category_id','tbl_category.id')
+                //                 ->where('request_momTask.id',$g)
+                //                 ->first();
+                $getSubmissionData = DB::table('request_mom')->where('id', $g)->first();
+
+                $getCreator = User::findOrFail($getSubmissionData->user_id); //  get creator
+
+                $mailData = [
+                    "all" => 1,
+                    "action_id" => 0,
+                    "submission" => $getSubmissionData,
+                    "email" => $getCreator->email, // kirim kepada creator
+                    "fullname" => $getCreator->fullname,
+                    "message" => $this->mailMessage()['deadlineTaskReminder'],
+                    "remarks" => null,
+                ];
+
+                Mail::to($mailData['email'])->send(new SubmissionMail($mailData,$this->modulename,1));
+            }
+        }
+
+
+    }
+
     public function destroy($id)
     {
         try {
 
             $data = $this->model->findOrFail($id);
+
+            $getTaskBound = $this->taskbound->where('task_id',$data->task_id)
+            ->where('employee_id',$this->getEmployeeID()->id)
+            ->get();
+
+            if(count($getTaskBound) < 1) {
+                return response()->json(["status" => "error", "message" => $this->getMessage()['nothaveaccess']]);
+            }
+
             $data->delete();
 
             return response()->json(["status" => "success", "message" => $this->getMessage()['destroy']]);
